Move the extension to the string class
MJ prefered the extension be with the `string` extensions even though we are extending the `IEnumerable`.
