+++ src/Useful.Extensions/StringExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Text.RegularExpressions;
@@ -407,5 +408,77 @@ namespace Useful.Extensions
         }
 
         #endregion IsAllAlphaOrNumbers
+
+        #region String.Join
+
+        public static string Join(this IEnumerable<string> source, char separator)
+        {
+            var result = string.Join(separator, source);
+            return result;
+        }
+
+        public static string Join(this IEnumerable<string> source, string separator)
+        {
+            var result = string.Join(separator, source);
+            return result;
+        }
+
+        // public static string Join(this IEnumerable<string> source, char separator, int startIndex, int count)
+        // {
+        //     var result = string.Join(separator, source, startIndex, count);
+        //     return result;
+        // }
+        //
+        // public static string Join(this IEnumerable<string> source, string separator, int startIndex, int count)
+        // {
+        //     var result = string.Join(separator, source, startIndex, count);
+        //     return result;
+        // }
+        
+        // public static string Join(
+        //     this string[] source,
+        //     char separator)
+        // {
+        //     var result = string.Join(separator, source);
+        //     return result;
+        // }
+        //
+        // public static string Join(
+        //     this string[] source,
+        //     string separator)
+        // {
+        //     var result = string.Join(separator, source);
+        //     return result;
+        // }
+        //
+        // public static string Join(
+        //     this string[] source,
+        //     char separator,
+        //     int startIndex,
+        //     int count)
+        // {
+        //     var result = string.Join(
+        //         separator,
+        //         source,
+        //         startIndex,
+        //         count);
+        //     return result;
+        // }
+        //
+        // public static string Join(
+        //     this string[] source,
+        //     string separator,
+        //     int startIndex,
+        //     int count)
+        // {
+        //     var result = string.Join(
+        //         separator,
+        //         source,
+        //         startIndex,
+        //         count);
+        //     return result;
+        // }
+
+        #endregion String.Join
     }
 }
